add convertResultSetToStudents and extract this method from DMLServiceImpl
