+++ StudentService/src/main/java/org/example/data/Student.java
@@ -1,6 +1,16 @@
 package org.example.data;
 
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+import lombok.experimental.FieldNameConstants;
+
+@Setter
+@Getter
+@ToString
+@FieldNameConstants
 public class Student {
+
     private String name;
     private String secondName;
     private int age;
@@ -12,38 +22,5 @@ public class Student {
     }
 
     public Student() {
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSecondName() {
-        return secondName;
-    }
-
-    public void setSecondName(String secondName) {
-        this.secondName = secondName;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    @Override
-    public String toString() {
-        return "Student{" +
-                "name='" + name + '\'' +
-                ", secondName='" + secondName + '\'' +
-                ", age=" + age +
-                '}';
     }
 }
