+++ StudentService/src/main/java/org/example/service/StudentConverter.java
@@ -3,7 +3,23 @@ package org.example.service;
 import org.example.data.Student;
 import org.example.data.dto.StudentDto;
 
-// TODO Add abstraction
+import java.sql.ResultSet;
+import java.util.List;
+
 public interface StudentConverter {
-    Student convertToStudent(StudentDto studentDto);
+
+    /**
+     * Converts student DTO to Student
+     * @param studentDto StudentDto object
+     * @return Student
+     */
+    Student convertDTOToStudent(StudentDto studentDto);
+
+    /**
+     * Composes Student object from the database selection
+     * @param resultSet receives ResultSet from database
+     * @return Student
+     */
+    List<Student> convertResultSetToStudents(ResultSet resultSet);
 }
+
